requirements.txt
aiohttp==3.6.2
APScheduler==3.6.3
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.4.4
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
multidict==4.7.4
pycodestyle==2.5.0
pyee==7.0.1
python-dotenv==0.11.0
pytz==2019.3
requests==2.22.0
selenium==3.141.0
six==1.14.0
slackclient==2.5.0
slackeventsapi==2.1.0
soupsieve==1.9.5
tzlocal==2.0.0
urllib3==1.25.8
virtualenv==16.7.6
Werkzeug==1.0.0
yarl==1.4.2
